Minor fixes.

+++ core/federated/network/lf_socket_support.c
@@ -12,7 +12,7 @@
 
 static void handle_header_read(unsigned char* buffer, size_t* bytes_to_read, int* state);
 
-netdrv_t* initialize_netdrv(uint16_t my_federate_id, const char* federation_id) {
+netdrv_t* initialize_netdrv(int my_federate_id, const char* federation_id) {
   netdrv_t* drv = initialize_common_netdrv(my_federate_id, federation_id);
 
   // Initialize priv.
@@ -79,7 +79,6 @@ void create_connector(netdrv_t* drv) {
 
 int connect_to_netdrv(netdrv_t* drv) {
   socket_priv_t* priv = (socket_priv_t*)drv->priv;
-
   int ret =
       connect_to_socket(priv->socket_descriptor, priv->server_hostname, priv->server_port, priv->user_specified_port);
   return ret;
